+++ Kinvey/KinveyTests/SyncStoreTests.swift
@@ -5317,10 +5317,10 @@ class SyncStoreTests: StoreTestCase {
     func testSyncStoreDisabledDeltasetWithSync() {
         let store = DataStore<Person>.collection(.sync)
         
-        var initialCount = Int64(0)
+        var initialCount = 0
         do {
             if !useMockData {
-                initialCount = Int64(try! DataStore<Person>.collection(.network).count(options: nil).waitForResult(timeout: defaultTimeout).value())
+                initialCount = try! DataStore<Person>.collection(.network).count(options: nil).waitForResult(timeout: defaultTimeout).value()
             }
         }
         
@@ -6613,7 +6613,7 @@ class SyncStoreTests: StoreTestCase {
                     let cacheCount = self.store.cache?.count(query: nil)
                     XCTAssertNotNil(cacheCount)
                     if let cacheCount = cacheCount {
-                        XCTAssertEqual(Int64(cacheCount), persons.count)
+                        XCTAssertEqual(cacheCount, persons.count)
                     }
                 case .failure(let error):
                     XCTFail(error.localizedDescription)
@@ -6693,7 +6693,7 @@ class SyncStoreTests: StoreTestCase {
                     let cacheCount = self.store.cache?.count(query: nil)
                     XCTAssertNotNil(cacheCount)
                     if let cacheCount = cacheCount {
-                        XCTAssertEqual(Int64(cacheCount), persons.count)
+                        XCTAssertEqual(cacheCount, persons.count)
                     }
                 case .failure(let error):
                     XCTFail(error.localizedDescription)
@@ -6754,7 +6754,7 @@ class SyncStoreTests: StoreTestCase {
                     let cacheCount = self.store.cache?.count(query: nil)
                     XCTAssertNotNil(cacheCount)
                     if let cacheCount = cacheCount {
-                        XCTAssertEqual(Int64(cacheCount), persons.count)
+                        XCTAssertEqual(cacheCount, persons.count)
                     }
                 case .failure(let error):
                     XCTFail(error.localizedDescription)
@@ -6868,7 +6868,7 @@ class SyncStoreTests: StoreTestCase {
                     let cacheCount = self.store.cache?.count(query: nil)
                     XCTAssertNotNil(cacheCount)
                     if let cacheCount = cacheCount {
-                        XCTAssertEqual(Int64(cacheCount), persons.count)
+                        XCTAssertEqual(cacheCount, persons.count)
                     }
                 case .failure(let error):
                     XCTFail(error.localizedDescription)
@@ -6929,7 +6929,7 @@ class SyncStoreTests: StoreTestCase {
                     let cacheCount = self.store.cache?.count(query: nil)
                     XCTAssertNotNil(cacheCount)
                     if let cacheCount = cacheCount {
-                        XCTAssertEqual(Int64(cacheCount), persons.count)
+                        XCTAssertEqual(cacheCount, persons.count)
                     }
                 case .failure(let error):
                     XCTFail(error.localizedDescription)
